Start moving functions that copy files into the image into a new install module
