+++ mkosi/gentoo.py
@@ -11,7 +11,7 @@ from pathlib import Path
 from textwrap import dedent
 from typing import Dict, Generator, List, Sequence
 
-from . import copy_path, open_close, unlink_try_hard
+from . import unlink_try_hard
 from .backend import (
     ARG_DEBUG,
     MkosiConfig,
@@ -25,6 +25,7 @@ from .backend import (
     run_workspace_command,
     safe_tar_extract,
 )
+from .install import copy_path, open_close
 
 ARCHITECTURES = {
     "x86_64": ("amd64", "arch/x86/boot/bzImage"),
